requirements.txt
torch>=2.0.0
transformers>=4.30.0
scanpy>=1.9.0
anndata>=0.8.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
toml>=0.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
